# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: User Management Control Panel API
  version: 0.1.0
  description: |
    Control-plane API for the user-management + licensing system.

    Covers admin authentication, organization/user/RBAC management, plan & subscription
    administration, license JWKS publication, key rotation, usage ingestion, audit log
    access, SSO configuration, API key management, and the domain event stream.

    Two authentication schemes are supported:
      - **bearerAuth**: short-lived admin session JWT, issued by `POST /api/v1/auth/login`.
      - **apiKeyAuth**: long-lived programmatic key, sent as `Authorization: ApiKey <token>`.

    The public `/.well-known/jwks.json` endpoint is unauthenticated.

    Error responses follow [RFC 7807 Problem Details](https://datatracker.ietf.org/doc/html/rfc7807).

    > **Caveat:** the license-issuance endpoint (`POST .../licenses`) referenced in the
    > implementation plan is *not yet wired up* in `control-panel-api`. It is documented
    > here as the intended contract for Phase 1 closure.
  contact:
    name: Control Panel Team
    email: support@control-panel.example.com
  license:
    name: Proprietary
servers:
  - url: https://control-panel.example.com
    description: Production
  - url: https://staging.control-panel.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local dev
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
    get:
      tags: [users]
      summary: Get a user.
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [users]
      summary: Update a user's profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/v1/users/{userId}/deactivate:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
    post:
      tags: [users]
      summary: Deactivate a user.
      responses:
        '204':
          description: Deactivated.
  /api/v1/users/{userId}/change-password:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
    post:
      tags: [users]
      summary: Change own password (the caller must equal `{userId}`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword: { type: string, format: password }
                newPassword: { type: string, format: password, minLength: 8 }
      responses:
        '204':
          description: Password changed.
        '403':
          $ref: '#/components/responses/Forbidden'

  # ============================================================
  # rbac
  # ============================================================
...
  /api/v1/orgs/{orgId}/subscriptions:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    get:
      tags: [subscriptions]
      summary: List subscriptions for an organization.
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
    post:
      tags: [subscriptions]
      summary: Create a subscription under an organization.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionCreate'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
  /api/v1/subscriptions/{subscriptionId}:
    parameters:
      - $ref: '#/components/parameters/SubscriptionIdParam'
    get:
      tags: [subscriptions]
      summary: Get a subscription.
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
  /api/v1/subscriptions/{subscriptionId}/suspend:
    parameters:
      - $ref: '#/components/parameters/SubscriptionIdParam'
    post:
      tags: [subscriptions]
      summary: Suspend an active subscription.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
  /api/v1/subscriptions/{subscriptionId}/reactivate:
    parameters:
      - $ref: '#/components/parameters/SubscriptionIdParam'
    post:
      tags: [subscriptions]
      summary: Reactivate a suspended subscription.
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
  /api/v1/subscriptions/{subscriptionId}/cancel:
    parameters:
      - $ref: '#/components/parameters/SubscriptionIdParam'
    post:
      tags: [subscriptions]
      summary: Cancel a subscription.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
  /api/v1/subscriptions/{subscriptionId}/overrides:
    parameters:
      - $ref: '#/components/parameters/SubscriptionIdParam'
    post:
      tags: [subscriptions]
      summary: Add a per-subscription override (permission add/remove or feature delta).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionOverride'
      responses:
        '200':
          description: Added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionOverride'
  /api/v1/subscriptions/{subscriptionId}/overrides/{overrideId}:
    parameters:
      - $ref: '#/components/parameters/SubscriptionIdParam'
      - in: path
        name: overrideId
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [subscriptions]
      summary: Remove an override.
      responses:
        '204':
          description: Removed.

  # ============================================================
  # licenses (PLANNED — see info note)
  # ============================================================
...
  /api/v1/usage/ingest:
    post:
      tags: [usage]
      summary: Ingest a batch of usage events tied to a license.
      description: |
        Authenticated. The body's `jti` identifies the license that generated the events;
        the server resolves the subscription from it and de-duplicates by `(jti, event_id)`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsageIngestRequest'
      responses:
        '202':
          description: Accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageIngestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/subscriptions/{subscriptionId}/usage:
    parameters:
      - $ref: '#/components/parameters/SubscriptionIdParam'
    get:
      tags: [usage]
      summary: Get events + quota rollups for a subscription.
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: featureKey
          schema: { type: string }
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageReport'

  # ============================================================
  # audit
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin session JWT from `POST /api/v1/auth/login`.
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Long-lived programmatic key.
        Format: `Authorization: ApiKey <token>` where `<token>` is the value returned
        once by `POST /api/v1/orgs/{orgId}/api-keys`.
  parameters:
    OrgIdParam:
      in: path
      name: orgId
      required: true
      schema: { type: string, format: uuid }
    UserIdParam:
      in: path
      name: userId
      required: true
      schema: { type: string, format: uuid }
    SubscriptionIdParam:
      in: path
      name: subscriptionId
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Request body or parameters invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    Forbidden:
      description: Authenticated but not authorized.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource does not exist or is hidden from this caller.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    ProblemDetail:
      type: object
      description: RFC 7807 problem details.
      properties:
        type: { type: string, format: uri, example: 'about:blank' }
        title: { type: string, example: 'Bad Request' }
        status: { type: integer, example: 400 }
        detail: { type: string, example: 'Field `email` is required.' }
        instance: { type: string, format: uri }
        code: { type: string, example: 'validation_failed' }
      required: [title, status]
    ReasonRequest:
      type: object
      properties:
        reason: { type: string }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        status:
          type: string
          enum: [ACTIVE, INVITED, DISABLED, DELETED]
        superAdmin: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time, nullable: true }
      required: [id, email, status]
    UserUpdate:
      type: object
      properties:
        fullName: { type: string, maxLength: 255 }

    # ---- org ----
...
    Subscription:
      type: object
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        planId: { type: string, format: uuid }
        status:
          $ref: '#/components/schemas/SubscriptionStatus'
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        seats: { type: integer }
        notes: { type: string }
        overrides:
          type: array
          items: { $ref: '#/components/schemas/SubscriptionOverrideRecord' }
      required: [id, orgId, planId, status]
    SubscriptionStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, EXPIRED, CANCELLED]
    SubscriptionCreate:
      type: object
      required: [planId]
      properties:
        planId: { type: string, format: uuid }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        seats: { type: integer, minimum: 0 }
        notes: { type: string }
        overrides:
          type: array
          items: { $ref: '#/components/schemas/SubscriptionOverride' }
    SubscriptionOverride:
      type: object
      required: [type, key]
      properties:
        type:
          type: string
          enum: [PERMISSION, FEATURE]
        key: { type: string }
        op:
          type: string
          enum: [ADD, REMOVE, SET]
          description: Defaults vary by type; PERMISSION uses ADD/REMOVE, FEATURE uses SET.
        value:
          description: For FEATURE/SET. Boolean/integer/string.
          oneOf:
            - type: boolean
            - type: integer
            - type: string
    SubscriptionOverrideRecord:
      type: object
      properties:
        id: { type: string, format: uuid }
        type:
          type: string
          enum: [PERMISSION, FEATURE]
        key: { type: string }
        value:
          oneOf:
            - type: boolean
            - type: integer
            - type: string
            - type: 'null'

    # ---- licenses ----
...
    UsageEvent:
      type: object
      required: [featureKey, quantity, occurredAt]
      properties:
        featureKey: { type: string, example: 'export.pdf' }
        quantity:
          type: number
          minimum: 0
        occurredAt: { type: string, format: date-time }
        metadata:
          type: object
          additionalProperties: true
    UsageIngestRequest:
      type: object
      required: [jti, events]
      properties:
        jti: { type: string }
        events:
          type: array
          items: { $ref: '#/components/schemas/UsageEvent' }
    UsageIngestResponse:
      type: object
      properties:
        eventsAccepted: { type: integer }
        subscriptionId: { type: string, format: uuid }
    UsageQuota:
      type: object
      properties:
        featureKey: { type: string }
        periodStart: { type: string, format: date-time }
        periodEnd: { type: string, format: date-time }
        limit:
          type: number
          nullable: true
        consumed:
          type: number
    UsageReport:
      type: object
      properties:
        events:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              featureKey: { type: string }
              quantity: { type: number }
              occurredAt: { type: string, format: date-time }
              jti: { type: string }
        quotas:
          type: array
          items: { $ref: '#/components/schemas/UsageQuota' }

    # ---- audit ----
...